+++ apps/main-app/ts/features/pn/hooks/__tests__/usePnPreferencesFetcher.test.tsx
@@ -1,4 +1,5 @@
 import * as pot from "@pagopa/ts-commons/lib/pot";
+import { useState } from "react";
 import { createStore } from "redux";
 import { ServiceId } from "../../../../../definitions/services/ServiceId";
 import { applicationChangeState } from "../../../../store/actions/application";
@@ -26,8 +27,6 @@ jest.mock("react", () => ({
   ...jest.requireActual("react"),
   useState: jest.fn()
 }));
-// eslint-disable-next-line import/order
-import { useState } from "react";
 
 type PreferencePotState = pot.Pot<
   ServicePreferenceResponse,