+++ apps/main-app/CHANGELOG.md
@@ -2,6 +2,77 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+## [3.36.0-rc.12](https://github.com/pagopa/io-app/compare/3.36.0-rc.11...3.36.0-rc.12) (2026-07-10)
+
+## [3.36.0-rc.11](https://github.com/pagopa/io-app/compare/3.36.0-rc.10...3.36.0-rc.11) (2026-07-09)
+
+
+### Features
+
+* **IT-Wallet:** [[SIW-4384](https://pagopa.atlassian.net/browse/SIW-4384),[SIW-4483](https://pagopa.atlassian.net/browse/SIW-4483)] Add Qualtrics survey entry points for eID and credential flows ([#8170](https://github.com/pagopa/io-app/issues/8170)) ([0e92db0](https://github.com/pagopa/io-app/commit/0e92db0f9f5131449090862131fd7d450982a833))
+* **IT-Wallet:** [[SIW-4621](https://pagopa.atlassian.net/browse/SIW-4621)] Add Status List repository ([#8255](https://github.com/pagopa/io-app/issues/8255)) ([69f2bb5](https://github.com/pagopa/io-app/commit/69f2bb532bc015ded6b70a9789b84e4ede84ed58))
+
+
+### Bug Fixes
+
+* **IT-Wallet:** [[SIW-4616](https://pagopa.atlassian.net/browse/SIW-4616)] Display credential intro screen before issuance  ([#8247](https://github.com/pagopa/io-app/issues/8247)) ([fd8f397](https://github.com/pagopa/io-app/commit/fd8f397bd269996bc00b5110398fcb295ae7111d))
+
+
+### Chores
+
+* [[IEL-684](https://pagopa.atlassian.net/browse/IEL-684)]] Fci's retry now check signature request status ([#8251](https://github.com/pagopa/io-app/issues/8251)) ([2a05114](https://github.com/pagopa/io-app/commit/2a0511413a8608f127debead5624c48bc20c4841))
+* **IT-Wallet:** [[SIW-0000](https://pagopa.atlassian.net/browse/SIW-0000)] update @pagopa/io-react-native-wallet to version 3.6.2 ([#8259](https://github.com/pagopa/io-app/issues/8259)) ([f36f1e1](https://github.com/pagopa/io-app/commit/f36f1e13fbca93b31787e1922ecd2d3bad967e57))
+* **IT-Wallet:** [[SIW-4427](https://pagopa.atlassian.net/browse/SIW-4427)] Enhance IT-Wallet upgrade flow steps ([#8172](https://github.com/pagopa/io-app/issues/8172)) ([df1cc4a](https://github.com/pagopa/io-app/commit/df1cc4a3a76dd1487d6247a3a5278db7943074da))
+* [[IOPLT-1947](https://pagopa.atlassian.net/browse/IOPLT-1947)] Enforce ESLint `no-non-null-assertion` rule as error ([#8250](https://github.com/pagopa/io-app/issues/8250)) ([7fde216](https://github.com/pagopa/io-app/commit/7fde216f26332fe1d9ad97ccd836528559d6694e))
+
+## [3.36.0-rc.10](https://github.com/pagopa/io-app/compare/3.36.0-rc.9...3.36.0-rc.10) (2026-07-08)
+
+
+### Features
+
+* [[IOCOM-3240](https://pagopa.atlassian.net/browse/IOCOM-3240)] addition of FSE banner  ([#8252](https://github.com/pagopa/io-app/issues/8252)) ([dfba458](https://github.com/pagopa/io-app/commit/dfba458498e535d9f02f6a35139829c919318bd0))
+* **IT-Wallet:** [[SIW-4475](https://pagopa.atlassian.net/browse/SIW-4475)] Update tour guide steps copy ([#8209](https://github.com/pagopa/io-app/issues/8209)) ([aad67c0](https://github.com/pagopa/io-app/commit/aad67c095b7aefb82aab85fb0331421ffb058032))
+
+
+### Bug Fixes
+
+* **IT-Wallet:** [[SIW-0000](https://pagopa.atlassian.net/browse/SIW-0000)] Update re-activation banner navigation to use `navigateToDiscoveryScreen` ([#8240](https://github.com/pagopa/io-app/issues/8240)) ([30ca466](https://github.com/pagopa/io-app/commit/30ca466ccea891cc3884a5489973f420d5a962f2))
+
+
+### Chores
+
+* **IT-Wallet:** [[SIW-000](https://pagopa.atlassian.net/browse/SIW-000)] Disable `usePreventScreenCapture` ([#8257](https://github.com/pagopa/io-app/issues/8257)) ([19b4370](https://github.com/pagopa/io-app/commit/19b4370b1c880a7bc38f45728b9bf91b77321ce7))
+* [[IOPLT-1943](https://pagopa.atlassian.net/browse/IOPLT-1943)] Enforce ESLint `no-unused-vars` rule as error ([#8241](https://github.com/pagopa/io-app/issues/8241)) ([bbd1df6](https://github.com/pagopa/io-app/commit/bbd1df6d89588123247b2b25d29e865e288d3cde))
+* [[IOPLT-1944](https://pagopa.atlassian.net/browse/IOPLT-1944)] Enforce ESLint `no-inferrable-types` rule as error ([#8242](https://github.com/pagopa/io-app/issues/8242)) ([ae156d2](https://github.com/pagopa/io-app/commit/ae156d25bbb2577c3c33978cc6a49aec2607a555))
+* [[IOPLT-1945](https://pagopa.atlassian.net/browse/IOPLT-1945)] Enforce ESLint `no-empty-function` rule as error ([#8245](https://github.com/pagopa/io-app/issues/8245)) ([9c5c789](https://github.com/pagopa/io-app/commit/9c5c7890070c0ec042ee2ea39021196b52b27e67))
+
+## [3.36.0-rc.9](https://github.com/pagopa/io-app/compare/3.36.0-rc.8...3.36.0-rc.9) (2026-07-07)
+
+
+### Features
+
+* **IT-Wallet:** [[SIW-4551](https://pagopa.atlassian.net/browse/SIW-4551)] Align credential offer intro screen to catalogue intro screen ([#8243](https://github.com/pagopa/io-app/issues/8243)) ([7cce544](https://github.com/pagopa/io-app/commit/7cce5449aa85565b7088b482575a00c2a689243c))
+
+## [3.36.0-rc.8](https://github.com/pagopa/io-app/compare/3.36.0-rc.7...3.36.0-rc.8) (2026-07-07)
+
+
+### Features
+
+* **IT-Wallet:** [[SIW-4084](https://pagopa.atlassian.net/browse/SIW-4084)] Implement `expo-background-task` for Status List background fetch ([#8233](https://github.com/pagopa/io-app/issues/8233)) ([ac5a4bc](https://github.com/pagopa/io-app/commit/ac5a4bc331bd4fbe1e471bcbec2b726e2182bb1b)), closes [/#diff-7ae45ad102eab3b6d7e7896acd08c427a9b25b346470d7bc6507b6481575d519R96-R98](https://github.com/pagopa///issues/diff-7ae45ad102eab3b6d7e7896acd08c427a9b25b346470d7bc6507b6481575d519R96-R98) [/#diff-e70329c240cbde250c4cf9499d16af44da3323edee966f9c4c2c55385da47008R137-R143](https://github.com/pagopa///issues/diff-e70329c240cbde250c4cf9499d16af44da3323edee966f9c4c2c55385da47008R137-R143) [/#diff-6701f6e68eb818df064691dfe50a2070f5af1dd45c5299c8299782d934a668f3R525](https://github.com/pagopa///issues/diff-6701f6e68eb818df064691dfe50a2070f5af1dd45c5299c8299782d934a668f3R525) [/#diff-6701f6e68eb818df064691dfe50a2070f5af1dd45c5299c8299782d934a668f3R544](https://github.com/pagopa///issues/diff-6701f6e68eb818df064691dfe50a2070f5af1dd45c5299c8299782d934a668f3R544) [/#diff-1679e0f2fa3191277b8e23ec38dd7abbe0543d528050a026d1580c242239ee31L16-R39](https://github.com/pagopa///issues/diff-1679e0f2fa3191277b8e23ec38dd7abbe0543d528050a026d1580c242239ee31L16-R39) [/#diff-1679e0f2fa3191277b8e23ec38dd7abbe0543d528050a026d1580c242239ee31R100-R103](https://github.com/pagopa///issues/diff-1679e0f2fa3191277b8e23ec38dd7abbe0543d528050a026d1580c242239ee31R100-R103) [/#diff-0b5810437eae8d523f9eb8c78aa8cb99d0034002e508c44e577badeb12c64034R1-R52](https://github.com/pagopa///issues/diff-0b5810437eae8d523f9eb8c78aa8cb99d0034002e508c44e577badeb12c64034R1-R52) [/#diff-e151b993d63bb2047f84910631be5f4976d0649c92d3cbb51b57fa3b1d1150faL47-R51](https://github.com/pagopa///issues/diff-e151b993d63bb2047f84910631be5f4976d0649c92d3cbb51b57fa3b1d1150faL47-R51) [/#diff-f3ba0db1a86822542f174ba66345e5b92234c0b90117445fbc0be8c16985be35R1-R26](https://github.com/pagopa///issues/diff-f3ba0db1a86822542f174ba66345e5b92234c0b90117445fbc0be8c16985be35R1-R26) [/#diff-c03472a911a4d4c33a562ca97db11c2d14896c4627003956c1579f90c1d57d7dR1-R132](https://github.com/pagopa///issues/diff-c03472a911a4d4c33a562ca97db11c2d14896c4627003956c1579f90c1d57d7dR1-R132) [/#diff-99b8b26eca513046f01a9bc3ffb95e710cdfab222fc96582b4cee396f9f0bea6L8-R14](https://github.com/pagopa///issues/diff-99b8b26eca513046f01a9bc3ffb95e710cdfab222fc96582b4cee396f9f0bea6L8-R14) [/#diff-99b8b26eca513046f01a9bc3ffb95e710cdfab222fc96582b4cee396f9f0bea6R28-R39](https://github.com/pagopa///issues/diff-99b8b26eca513046f01a9bc3ffb95e710cdfab222fc96582b4cee396f9f0bea6R28-R39) [/#diff-99b8b26eca513046f01a9bc3ffb95e710cdfab222fc96582b4cee396f9f0bea6L45-R119](https://github.com/pagopa///issues/diff-99b8b26eca513046f01a9bc3ffb95e710cdfab222fc96582b4cee396f9f0bea6L45-R119)
+
+## [3.36.0-rc.7](https://github.com/pagopa/io-app/compare/3.36.0-rc.6...3.36.0-rc.7) (2026-07-06)
+
+
+### Bug Fixes
+
+* **IT-Wallet:** [[SIW-0000](https://pagopa.atlassian.net/browse/SIW-0000)] Credentials catalog, claims display, mdoc verification ([#8236](https://github.com/pagopa/io-app/issues/8236)) ([bf8a20d](https://github.com/pagopa/io-app/commit/bf8a20d34e8e8ee551c76451e27015d1ac21e04a))
+
+
+### Chores
+
+* **IT-Wallet:** [[SIW-3612](https://pagopa.atlassian.net/browse/SIW-3612)] Display proximity CTA in all mdoc credentials detail screen ([#8237](https://github.com/pagopa/io-app/issues/8237)) ([6844f51](https://github.com/pagopa/io-app/commit/6844f51196c8f88213b597b42db58f43cb78c2e7))
+* [[IOPLT-1751](https://pagopa.atlassian.net/browse/IOPLT-1751)] Add `@pagopa/eslint-config` to the `eslint` configuration ([#7992](https://github.com/pagopa/io-app/issues/7992)) ([a02a0f6](https://github.com/pagopa/io-app/commit/a02a0f68e2ca048d9b25dfa3bb02edda6c8c7276))
+
 ## [3.36.0-rc.6](https://github.com/pagopa/io-app/compare/3.36.0-rc.5...3.36.0-rc.6) (2026-07-03)